chore: make framework selection based on dependencies
