+++ src/utils/vscode-utils.ts
@@ -1,8 +1,14 @@
-import { window, workspace } from 'vscode';
-import { MessageTypes } from './message-types';
 import * as fs from 'fs';
 import * as path from 'path';
+import { window, workspace } from 'vscode';
+import * as constants from '../constants';
+import { FileService } from './file-utils';
+import { MessageTypes } from './message-types';
 
+interface PackageJson {
+	dependencies: [];
+	devdependencies: [];
+}
 export class VSCodeService {
     private static _instance: VSCodeService;
 
@@ -37,7 +43,7 @@ export class VSCodeService {
             return;
         }
 
-        var files = fs.readdirSync(startPath).filter(e => e.indexOf('node_modules'))
+        var files = fs.readdirSync(startPath).filter(e => e.indexOf('node_modules'));
         for (var i = 0; i < files.length; i++) {
             var filename = path.join(startPath, files[i]);
             var stat = fs.lstatSync(filename);
@@ -46,9 +52,9 @@ export class VSCodeService {
             }
             else if (filename.indexOf(filter) >= 0) {
                 fileArray.push(filename);
-            };
-        };
-    };
+            }
+        }
+    }
 
     async getFilesByPattern(folderPath: string, pattern: string = '.sketch') {
         let fileArray: string[] = [];
@@ -58,5 +64,25 @@ export class VSCodeService {
 
     async showQuickDialog(items: string[], placeHolder: string, multiSelection: boolean): Promise<any> {
         return await window.showQuickPick(items, { placeHolder, matchOnDetail: true, canPickMany: multiSelection });
+    }
+
+    public getFrameworksFromWorkspace = () => {
+        const fileUtils = FileService.getInstance();
+        const workspacePath = workspace.workspaceFolders && workspace.workspaceFolders[0];
+        const packagePath = `${workspacePath!.uri.fsPath}/package.json`;
+        const packageExist = fileUtils.fileExists(packagePath);
+        if (packageExist) {
+            const bufferdPackage = fileUtils.readFile(packagePath);
+            const packageStrinbg = JSON.parse(bufferdPackage.toString()) as PackageJson;
+            return this.findFrameworkPackages(packageStrinbg);
+        }
+        return [];
+    
+    }
+    
+    private findFrameworkPackages = (packageJson: PackageJson) => {
+        const packages = Object.keys(packageJson.dependencies);
+        return constants.FRAMEWORKS.filter(framework => packages.find(e => e.includes(framework)));
+    
     }
 }
